+++ apps/extension/src/ui/domains/Transactions/TxHistory/TxHistoryDetails/TxHistoryDetailsTokens.tsx
@@ -1,0 +1,10 @@
+import { FC } from "react"
+
+import { TokensAndFiat } from "@ui/domains/Asset/TokensAndFiat"
+
+export const TxHistoryDetailsTokens: FC<{
+  value: bigint | string
+  tokenId: string
+}> = ({ value, tokenId }) => {
+  return <TokensAndFiat planck={value} tokenId={tokenId} />
+}