+++ packages/extension-core/src/domains/app/store.remoteConfig.ts
@@ -37,7 +37,6 @@ export const DEFAULT_REMOTE_CONFIG: RemoteConfigStoreData = {
     // uncomment for testing on testnets
     // "avail-turing-testnet": [1],
     // "vara-testnet": [1],
-    // "aleph-zero-testnet": [1],
   },
   stakingPools: {},
   documentation: {
@@ -76,6 +75,12 @@ export const DEFAULT_REMOTE_CONFIG: RemoteConfigStoreData = {
       },
     ],
   },
+  bittensor: {
+    fee: {
+      buy: {},
+      sell: {},
+    },
+  },
 }
 
 const CONFIG_TIMEOUT = 30 * 60 * 1000 // 30 minutes
@@ -100,6 +105,7 @@ export class RemoteConfigStore extends StorageProvider<RemoteConfigStoreData> {
           config.featureFlags.SEEK_BENEFITS = true
           config.featureFlags.SEEK_TAO_DISCOUNT = true
           config.featureFlags.ASSET_HUB_MIGRATION_BANNER = true
+          config.featureFlags.SEEK_PRESALE = true
         }
 
         // first arg is an empty object so that DEFAULT_REMOTE_CONFIG is not mutated