+++ packages/on-chain-id/src/util/isPotentialEns.ts
@@ -1,5 +1,2 @@
-import { isPotentialAzns } from "./isPotentialAzns"
-
-/** dot separated string e.g. `ens.eth` or `hello.lol`, but not ending with `.azero` nor `.tzero` */
-export const isPotentialEns = (name?: string) =>
-  typeof name === "string" && /^.+\..+$/.test(name) && !isPotentialAzns(name)
+/** dot separated string e.g. `ens.eth` or `hello.lol` */
+export const isPotentialEns = (name?: string) => typeof name === "string" && /^.+\..+$/.test(name)